requirements.txt
# Core functionality
python-dotenv==1.0.1  # Environment variables
pynput==1.7.6  # Keyboard shortcuts
sounddevice==0.4.6  # Audio recording
soundfile==0.12.1  # Audio file handling

openai==1.51.2
anthropic==0.36.1

# UI and system interaction
pyautogui==0.9.54  # Cursor/text manipulation
pystray==0.19.5  # System tray icon
Pillow==10.1.0  # Required by pystray for icons
numpy==1.26.4
pyperclip==1.9.0